% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/united.R
\docType{data}
\name{A2Benchmark}
\alias{A2Benchmark}
\title{Yahoo Webscope S5 – A2 Benchmark (Synthetic)}
\format{
A list of time series.
}
\source{
\doi{10.1371/journal.pone.0262463}
}
\usage{
data(A2Benchmark)
}
\description{
Part of the Yahoo Webscope S5 dataset. A2 contains synthetic
time series with labeled anomalies designed to stress-test algorithms.
Labels available: Yes.
}
\details{
This package ships a mini version of the dataset. Use loadfulldata() to download and load the full dataset from the URL stored in attr(url).
}
\examples{
data(A2Benchmark)
s <- A2Benchmark[[1]]
summary(s$value)
}
\references{
Yoshihara K, Takahashi K (2022) A simple method for unsupervised anomaly detection: An application to Web time series data. PLoS ONE 17(1).

Chandola, V., Banerjee, A., & Kumar, V. (2009). Anomaly detection: A survey. ACM Computing Surveys, 41(3), 1–58.
}
\keyword{datasets}
