% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/united.R
\docType{data}
\name{A3Benchmark}
\alias{A3Benchmark}
\title{Yahoo Webscope S5 – A3 Benchmark (Synthetic with Outliers)}
\format{
A list of time series.
}
\source{
\doi{10.1371/journal.pone.0262463}
}
\usage{
data(A3Benchmark)
}
\description{
Part of the Yahoo Webscope S5 dataset. A3 contains synthetic
time series with labeled outliers/anomalies. Labels available: Yes.
}
\details{
This package ships a mini version of the dataset. Use loadfulldata() to download and load the full dataset from the URL stored in attr(url).
}
\examples{
library(harbinger)
data(A3Benchmark)
s <- A3Benchmark[[1]]
# Quick visualization with harbinger
har_plot(harbinger(), s$value)
}
\references{
Yoshihara K, Takahashi K (2022) A simple method for unsupervised anomaly detection: An application to Web time series data. PLoS ONE 17(1).
}
\keyword{datasets}
