% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/united.R
\docType{data}
\name{A4Benchmark}
\alias{A4Benchmark}
\title{Yahoo Webscope S5 – A4 Benchmark (Synthetic with Anomalies and CPs)}
\format{
A list of time series.
}
\source{
\doi{10.1371/journal.pone.0262463}
}
\usage{
data(A4Benchmark)
}
\description{
Part of the Yahoo Webscope S5 dataset. A4 contains synthetic
time series with labeled anomalies and change points. Labels available: Yes.
}
\details{
This package ships a mini version of the dataset. Use loadfulldata() to download and load the full dataset from the URL stored in attr(url).
}
\examples{
data(A4Benchmark)
s <- A4Benchmark[[1]]
mean(s$event)  # proportion of anomalous or change-point timestamps
}
\references{
Yoshihara K, Takahashi K (2022) A simple method for unsupervised anomaly detection: An application to Web time series data. PLoS ONE 17(1).

Truong, C., Oudre, L., & Vayatis, N. (2020). Selective review of change point detection methods. Signal Processing, 167, 107299.
}
\keyword{datasets}
