% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cla_dtree.R
\name{cla_dtree}
\alias{cla_dtree}
\title{Decision Tree Classification}
\usage{
cla_dtree(attribute, slevels = NULL)
}
\arguments{
\item{attribute}{\itemize{
\item name of the attribute used as target classification
}}

\item{slevels}{\itemize{
\item possible values for the target classification
}}
}
\value{
classification object
}
\description{
Classification using Decision Tree algorithm
}
