% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cla_knn.R
\name{cla_knn}
\alias{cla_knn}
\title{K Nearest Neighbor Classification}
\usage{
cla_knn(attribute, slevels = NULL, k = 1:30)
}
\arguments{
\item{attribute}{\itemize{
\item name of the attribute used as target classification
}}

\item{slevels}{\itemize{
\item possible values for the target classification
}}

\item{k}{\itemize{
\item number od classes
}}
}
\value{
classification object
}
\description{
Classification using KNN algorithm
}
