% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cla_majority.R
\name{cla_majority}
\alias{cla_majority}
\title{Majority Classification}
\usage{
cla_majority(attribute, slevels = NULL)
}
\arguments{
\item{attribute}{\itemize{
\item name of the attribute used as target classification
}}

\item{slevels}{\itemize{
\item possible values for the target classification
}}
}
\value{
classification object
}
\description{
Majority Classification
}
