% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cla_nb.R
\name{cla_nb}
\alias{cla_nb}
\title{Naive Bayes Classifier}
\usage{
cla_nb(attribute, slevels = NULL)
}
\arguments{
\item{attribute}{\itemize{
\item name of the attribute used as target classification
}}

\item{slevels}{\itemize{
\item possible values for the target classification
}}
}
\description{
Naive Bayes Classifier
}
