% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cla_svm.R
\name{cla_svm}
\alias{cla_svm}
\title{Support Vector Machine Classification}
\usage{
cla_svm(
  attribute,
  slevels = NULL,
  epsilon = seq(0, 1, 0.2),
  cost = seq(20, 100, 20),
  kernel = "radial"
)
}
\arguments{
\item{attribute}{\itemize{
\item name of the attribute used as target classification
}}

\item{slevels}{\itemize{
\item possible values for the target classification
}}
}
\description{
Support Vector Machine Classification
}
