% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cla_main.R
\name{classification}
\alias{classification}
\title{classification class}
\usage{
classification(attribute, slevels = NULL)
}
\arguments{
\item{attribute}{\itemize{
\item name of the attribute used as target classification
}}

\item{slevels}{\itemize{
\item possible values for the target classification
}}
}
\value{
classification object
}
\description{
Ancestor class for classification problems
}
\details{
basic wrapper for classification problems
}
\examples{
data(iris)
template_model <- classification("Species", levels(iris$Species))
}
