% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/har_change_finder_arima.R
\name{detect.change_finder_arima}
\alias{detect.change_finder_arima}
\title{Detection of changes in time series using ARIMA models}
\usage{
\method{detect}{change_finder_arima}(obj, serie)
}
\arguments{
\item{serie}{}
}
\value{
A table indicating the positions where the outliers occur and changes in the time series
}
\description{
The function takes as parameters an object and a vector 'serie' containing the time series values that will be analized
}
\details{
First of all, the function adjusts an Arima model for all time series using the 'auto.arima' function from 'forecast' package. After of this, the function calculates the residuals of this model and uses 'outliers.boxplot.index' function from 'rugarch' package to identify outliers in this vector of residuals. Then, the function applies a sliding window of w size in the residuals, using the function 'TSPred::mas' to smooth the values
}
