% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/har_fbiad.R
\name{detect.fbiad}
\alias{detect.fbiad}
\title{Implements a time series anomaly detection method}
\usage{
\method{detect}{fbiad}(obj, serie)
}
\arguments{
\item{serie}{}
}
\value{
A detection table is returned with the index of each data point, its anomaly state (TRUE or FALSE) and an event type (defined as "anomaly" in this function)
}
\description{
Takes as input an object and a time series. It uses the FBIAD method to detect anomalies in the time series.
}
\details{
First, the function checks if there is data available in the time series. Then, the series is divided into windows of size 'obj$sw', and the average in each window is calculated. The quadratic difference between the original series and the mean of the window is then calculated and subjected to a boxplot-based anomaly detection test with threshold defined by 'obj$alpha'
}
