% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hanc_ml.R
\name{detect.hanc_ml}
\alias{detect.hanc_ml}
\title{Anomaly detector based on machine learning classification}
\usage{
\method{detect}{hanc_ml}(obj, serie, ...)
}
\arguments{
\item{obj}{detector}

\item{serie}{time series}

\item{...}{optional arguments.}
}
\value{
A dataframe with information about the detected anomalous points
}
\description{
Takes as input a "Harbinger" object and a time series
}
\examples{
detector <- harbinger()
}
