% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hanr_arima.R
\name{detect.hanr_arima}
\alias{detect.hanr_arima}
\title{Detect anomalies in time series using ARIMA models}
\usage{
\method{detect}{hanr_arima}(obj, serie, ...)
}
\arguments{
\item{obj}{detector}

\item{serie}{time series}

\item{...}{optional arguments.}
}
\value{
A dataframe with information about the detected anomalous points
}
\description{
Takes as input a "Harbinger" object and a time series
}
\examples{
detector <- harbinger()
}
