% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hanr_fbiad.R
\name{detect.hanr_fbiad}
\alias{detect.hanr_fbiad}
\title{Anomaly detector using Forward-Backward Inertia Anomaly Detection}
\usage{
\method{detect}{hanr_fbiad}(obj, serie, ...)
}
\arguments{
\item{obj}{detector}

\item{serie}{time series}

\item{...}{optional arguments.}
}
\value{
A dataframe with information about the detected anomalous points
}
\description{
Takes as input a "Harbinger" object and a time series
}
