% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hanr_kmeans.R
\name{detect.hanr_kmeans}
\alias{detect.hanr_kmeans}
\title{Anomaly detector using kmeans}
\usage{
\method{detect}{hanr_kmeans}(obj, serie, ...)
}
\arguments{
\item{obj}{detector}

\item{serie}{time series}

\item{...}{optional arguments.}
}
\value{
A dataframe with information about the detected anomalous points
}
\description{
Takes as input a "Harbinger" object and a time series
}
