% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/har_arima.R
\name{detect.har_arima}
\alias{detect.har_arima}
\title{Detect anomalies in time series using ARIMA models}
\usage{
\method{detect}{har_arima}(obj, serie)
}
\arguments{
\item{serie}{}
}
\value{
A dataframe with information about the detected anomalous points
}
\description{
Implements a specific method for detecting anomalies in time series using an ARIMA model fitted to the data provided as input
}
\details{
It uses the "forecast" R package to fit the ARIMA model and the "rugaarch" package to calculate model residuals. It then uses the "TSPred" package to identify outliers using a boxplot and marks the anomalous points in the resulting time series
}
