% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/har_tsreg_sw.R
\name{detect.har_tsreg_sw}
\alias{detect.har_tsreg_sw}
\title{Implements the detect method of the har_tsreg_sw class for detecting anomalies in time series}
\usage{
\method{detect}{har_tsreg_sw}(obj, serie)
}
\arguments{
\item{serie}{}
}
\value{
The i_outliers array that is included in the detection data structure
}
\description{
The function receives as parameter an object and a time series
}
\details{
The function starts by defining the length of the time series n and the index of non-missing observations non_na. The function then calls the ts_data function to create a time series object with sliding windows of size obj$sw_size and stores in io the array of input values and vector of output values for the model
}
