% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/harbinger.R
\name{detect.harbinger}
\alias{detect.harbinger}
\title{Generic detector}
\usage{
\method{detect}{harbinger}(obj, serie, ...)
}
\arguments{
\item{obj}{detector}

\item{serie}{time series}

\item{...}{optional arguments.}
}
\value{
A dataframe with information about the detected anomalous points
}
\description{
Takes as input a "Harbinger" object and a time series
}
\examples{
detector <- harbinger()
}
