% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hcp_cf_ets.R
\name{detect.hcp_cf_ets}
\alias{detect.hcp_cf_ets}
\title{Change Finder Using ETS}
\usage{
\method{detect}{hcp_cf_ets}(obj, serie, ...)
}
\arguments{
\item{obj}{detector}

\item{serie}{time series}

\item{...}{optional arguments.}
}
\value{
A dataframe with information about the detected anomalous points
}
\description{
Takes as input a "Harbinger" object and a time series
}
\examples{
detector <- harbinger()
}
