% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hcp_scp.R
\name{detect.hcp_scp}
\alias{detect.hcp_scp}
\title{Detect change points in time series}
\usage{
\method{detect}{hcp_scp}(obj, serie)
}
\arguments{
\item{serie}{}
}
\value{
The function returns a dataframe indicating the occurrence of change points at each point in the series
}
\description{
The function takes an object object and a serial time series as input
}
\details{
Detection is done by applying the analyze_window function to each sliding window of data and identifying points with significantly different fit error between the two halves of the window. These points are marked as change points
}
