% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hmo_sax.R
\name{detect.hmo_sax}
\alias{detect.hmo_sax}
\title{Motif discovery using SAX}
\usage{
\method{detect}{hmo_sax}(obj, serie, ...)
}
\arguments{
\item{obj}{detector}

\item{serie}{time series}

\item{...}{optional arguments.}
}
\value{
A dataframe with information about the detected anomalous points
}
\description{
Takes as input a "Harbinger" object and a time series
}
