% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hmu_pca.R
\name{detect.hmu_pca}
\alias{detect.hmu_pca}
\title{Multivariate anomaly detector using PCA}
\usage{
\method{detect}{hmu_pca}(obj, serie, ...)
}
\arguments{
\item{obj}{detector}

\item{serie}{time series}

\item{...}{optional arguments.}
}
\value{
A dataframe with information about the detected anomalous points
}
\description{
Multivariate anomaly detector using PCA
}
\examples{
detector <- harbinger()
}
