% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/harbinger.R
\name{evaluate.harbinger}
\alias{evaluate.harbinger}
\title{Evaluate the performance of the event detection model against true events}
\usage{
\method{evaluate}{harbinger}(obj, detection, event, evaluation = har_eval(), ...)
}
\arguments{
\item{obj}{detector}

\item{detection}{detected observations}

\item{event}{labeled events}

\item{evaluation}{evaluation object}

\item{...}{optional arguments.}
}
\value{
The result of the evaluation
}
\description{
Takes as input a "Harbinger" object, a set of event detection (stored as a data frame) and a set of true events (also stored as a data frame)
}
\examples{
detector <- harbinger()
}
