% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/har_hard_evaluation.R
\name{evaluate.hard_evaluation}
\alias{evaluate.hard_evaluation}
\title{Evaluates the model based on a confusion matrix}
\usage{
\method{evaluate}{hard_evaluation}(obj, detection, event)
}
\arguments{
\item{event}{}
}
\value{
An object of the "hard evaluation" class updated with the calculated metrics
}
\description{
It takes three arguments: obj: an object to store the evaluation results; detection: a boolean vector that indicates whether or not the model detected an event; event: a boolean vector that indicates whether the event occurred or not
}
\details{
The function calculates various detection model benchmark metrics based on the information provided in the detection and event vectors
}
