% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/har_hard_evaluation.R
\name{evaluate}
\alias{evaluate}
\title{Fit a model for event detection}
\usage{
evaluate(obj, ...)
}
\arguments{
\item{obj}{harbinger object}

\item{...}{optional arguments./ further arguments passed to or from other methods.}
}
\value{
a harbinger object with model details
}
\description{
Basic ancestor function for build model for event detection
}
\details{
The fit function builds a model for time series event detection.
For some methods, the model is not needed to be build, so the function do nothing.
}
\examples{
evaluation <- hard_evaluation()
evaluation <- evaluate(evaluation, c(1, 0, 1, 0), c(0, 0, 1, 0))
}
