% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/har_soft_evaluation.R
\name{evaluate.soft_evaluation}
\alias{evaluate.soft_evaluation}
\title{Evaluate the quality of an event detection model against a reference dataset}
\usage{
\method{evaluate}{soft_evaluation}(obj, detection, event)
}
\arguments{
\item{event}{}
}
\value{
A list of performance measures for detection
}
\description{
It receives as input an object, a "detection" matrix with the detections made by the model and an "event" matrix with the true events
}
\details{
The function converts the NA values of the "detection" array to FALSE. It then calls the "soft_scores" function to calculate soft scores (between 0 and 1) of correspondence between detections and true events
}
