% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{examples_changepoints}
\alias{examples_changepoints}
\title{Time series for change point detection}
\format{
A list of time series for change point detection.
}
\source{
\href{https://github.com/cefet-rj-dal/harbinger}{Harbinger package}
}
\usage{
data(examples_changepoints)
}
\description{
A list of time series for change point
\itemize{
\item simple: a simple synthetic time series
\item sinusoidal: a sinusoidal synthetic time series
\item incremental: a incremental synthetic time series
\item abrupt: a abrupt synthetic time series
\item volatility: a volatility synthetic time series
}
#'
}
\examples{
data(examples_changepoints)
serie <- examples_changepoints$simple
}
\references{
\href{https://github.com/cefet-rj-dal/harbinger}{Harbinger package}
}
\keyword{datasets}
