% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{examples_harbinger}
\alias{examples_harbinger}
\title{Time series for event detection}
\format{
A list of time series.
}
\source{
\href{https://github.com/cefet-rj-dal/harbinger}{Harbinger package}
}
\usage{
data(examples_harbinger)
}
\description{
A list of time series for event detection
\itemize{
\item nonstationarity: a synthetic nonstationarity time series
\item global_temperature_yearly: yearly global temperature of the world
\item global_temperature_monthly: monthly global temperature of the world
\item multidimensional: multidimensional time series with a change point
\item seattle_week: Seattle weakly temperature in 2019
\item seattle_daily: Seattle daily temperature in 2019
}
#'
}
\examples{
data(examples_harbinger)
serie <- examples_harbinger$seattle_daily
}
\references{
\href{https://github.com/cefet-rj-dal/harbinger}{Harbinger package}
}
\keyword{datasets}
