% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{examples_motifs}
\alias{examples_motifs}
\title{Time series for change point detection}
\format{
A list of time series for motif discovery.
}
\source{
\href{https://github.com/cefet-rj-dal/harbinger}{Harbinger package}
}
\usage{
data(examples_motifs)
}
\description{
A list of time series for change point
\itemize{
\item simple: a simple synthetic time series
\item mitdb100: sample of mitdb 100 time series
\item mitdb102: sample of mitdb 102 time series
}
#'
}
\examples{
data(examples_motifs)
serie <- examples_motifs$simple
}
\references{
\href{https://github.com/cefet-rj-dal/harbinger}{Harbinger package}
}
\keyword{datasets}
