% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hanr_ml.R
\name{fit.hanr_ml}
\alias{fit.hanr_ml}
\title{Implements the fit method of the hanr_ml class to fit a regression model on the time series passed as series argument}
\usage{
\method{fit}{hanr_ml}(obj, serie)
}
\arguments{
\item{serie}{}
}
\value{
The 'obj' object updated with the adjusted model
}
\description{
The function receives as parameter an object and a time series
}
\details{
The function starts by creating a time series object with the ts_data function, which divides the series into sliding windows of size obj$sw_size and returns a list with two columns: input and output
}
