% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/harbinger.R
\name{fit.harbinger}
\alias{fit.harbinger}
\title{Fit a general outline of a fit function}
\usage{
\method{fit}{harbinger}(obj, serie)
}
\arguments{
\item{serie}{}
}
\value{
The "Harbinger" object without doing any processing or tuning of the model
}
\description{
Takes as input a "Harbinger" object and a time series
}
\details{
This function is a default implementation of the fit() function for the "Harbinger" class, which can be overridden by a specific subclass to perform proper model fit
}
