% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hanr_arima.R
\name{hanr_arima}
\alias{hanr_arima}
\title{Anomaly detector using ARIMA}
\usage{
hanr_arima(w = NULL)
}
\arguments{
\item{w}{Window size for warm-up ARIMA}
}
\value{
hanr_arima object
}
\description{
Anomaly detector using ARIMA
}
\examples{
detector <- harbinger()
}
