% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hanr_fft_binseg_cusum.R
\name{hanr_fft_binseg_cusum}
\alias{hanr_fft_binseg_cusum}
\title{Anomaly Detector using FFT with BinSeg and CUSUM Cutoff}
\usage{
hanr_fft_binseg_cusum()
}
\value{
\code{hanr_fft_binseg_cusum} object
}
\description{
This detector combines FFT-based spectral filtering with a Binary Segmentation
change-point cutoff applied to the cumulative spectrum. The lower-frequency
components are removed, the signal is reconstructed, and the residual is scored
for anomalies.

This function extends the HARBINGER framework and returns an object of class \code{hanr_fft_binseg_cusum}.
}
\examples{
library(daltoolbox)

# Load anomaly example data
data(examples_anomalies)

#Using simple example
dataset <- examples_anomalies$simple
head(dataset)

# setting up time series fft detector
model <- hanr_fft_binseg_cusum()

# fitting the model
model <- fit(model, dataset$serie)

# Run detection
detection <- detect(model, dataset$serie)

# filtering detected events
print(detection[(detection$event),])

}
\references{
\itemize{
\item Sobrinho, E. P., Souza, J., Lima, J., Giusti, L., Bezerra, E., Coutinho, R., Baroni, L.,
Pacitti, E., Porto, F., Belloze, K., Ogasawara, E. Fine-Tuning Detection Criteria for
Enhancing Anomaly Detection in Time Series. In: Simpósio Brasileiro de Banco de Dados
(SBBD). SBC, 29 Sep. 2025. doi:10.5753/sbbd.2025.247063
}
}
