% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hanr_garch.R
\name{hanr_garch}
\alias{hanr_garch}
\title{Anomaly detector using Garch}
\usage{
hanr_garch(w = 5, alpha = 1.5)
}
\arguments{
\item{w}{Window size for warm-up garch}

\item{alpha}{Threshold for outliers}
}
\value{
hanr_garch object
}
\description{
Anomaly detector using Garch
}
\examples{
detector <- harbinger()
}
