% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hanr_kmeans.R
\name{hanr_kmeans}
\alias{hanr_kmeans}
\title{Anomaly detector using kmeans}
\usage{
hanr_kmeans(seq = 1, centers = NA, alpha = 1.5)
}
\arguments{
\item{seq}{Sequence size}

\item{centers}{Number of centroids}

\item{alpha}{Threshold for outliers}
}
\value{
hanr_kmeans object
}
\description{
Anomaly detector using kmeans
}
\examples{
detector <- harbinger()
}
