% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/har_eval_soft.R
\name{har_eval_soft}
\alias{har_eval_soft}
\title{Evaluation of event detection}
\usage{
har_eval_soft(sw_size = 15)
}
\arguments{
\item{sw_size}{tolerance window size}
}
\value{
\code{har_eval_soft} object
}
\description{
Evaluation of event detection using SoftED \url{doi:10.48550/arXiv.2304.00439}
}
\examples{
#loading the example database
data(har_examples)
}
