% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{har_examples}
\alias{har_examples}
\title{Synthetic data set for time series analysis
Each time series has 101 observations with a single event on it.}
\format{
A list of time series.
}
\source{
\href{https://github.com/cefet-rj-dal/harbinger}{Harbinger package}
}
\usage{
data(har_examples)
}
\description{
Synthetic data set for time series analysis
Each time series has 101 observations with a single event on it.
}
\examples{
data(har_examples)
serie <- har_examples[[1]]
}
\references{
\href{https://github.com/cefet-rj-dal/harbinger}{Harbinger package}
}
\keyword{datasets}
