% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/har_outliers.R
\name{har_outliers}
\alias{har_outliers}
\title{Detect outliers in a dataset using the boxplot method}
\usage{
har_outliers(data, alpha = 1.5)
}
\arguments{
\item{data}{dataset}

\item{alpha}{Threshold for outliers}
}
\value{
The function returns a boolean array with true values for values considered to be outliers
}
\description{
The function receives a dataset "data" and an "alpha" parameter (default 1.5), which controls the upper limit of the boxplot
}
\examples{
detector <- harbinger()
}
