% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/har_outliers.R
\name{har_outliers_idx}
\alias{har_outliers_idx}
\title{Detect and return the indices of outliers in a dataset using the boxplot method}
\usage{
har_outliers_idx(data, alpha = 1.5)
}
\arguments{
\item{data}{dataset}

\item{alpha}{Threshold for outliers}
}
\value{
A boolean vector indicating whether each value in the data set is an outlier or not and a vector of indices of the values considered to be outliers
}
\description{
The function receives a dataset "data" and an "alpha" parameter (default 1.5), which controls the upper limit of the boxplot
}
\examples{
detector <- harbinger()
}
