% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/har_online_sources.R
\name{har_source}
\alias{har_source}
\title{Base source object}
\usage{
har_source(name = "source")
}
\arguments{
\item{name}{Human-readable source name.}
}
\value{
A \code{har_source} object.
}
\description{
Creates the common base used by Harbinger streaming sources.
}
\keyword{internal}
