% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hcd_kswin.R
\name{hcd_kswin}
\alias{hcd_kswin}
\title{KSWIN method}
\usage{
hcd_kswin(window_size = 100, stat_size = 30, alpha = 0.005, data = NULL)
}
\arguments{
\item{window_size}{Size of the sliding window (must be > 2*stat_size)}

\item{stat_size}{Size of the statistic window}

\item{alpha}{Probability for the test statistic of the Kolmogorov-Smirnov-Test The alpha parameter is very sensitive, therefore should be set below 0.01.}

\item{data}{Already collected data to avoid cold start.}
}
\value{
\code{hcd_kswin} object
}
\description{
Kolmogorov-Smirnov Windowing method for concept drift detection \url{doi:10.1016/j.neucom.2019.11.111}.
}
\examples{
library("daltoolbox")
library("ggplot2")

n <- 60  # size of each segment
serie1 <- c(sin((1:n)/pi), 2*sin((1:n)/pi), 10 + sin((1:n)/pi),
           10-10/n*(1:n)+sin((1:n)/pi)/2, sin((1:n)/pi)/2)
serie2 <- 2*c(sin((1:n)/pi), 2*sin((1:n)/pi), 10 + sin((1:n)/pi),
             10-10/n*(1:n)+sin((1:n)/pi)/2, sin((1:n)/pi)/2)
data <- data.frame(serie1, serie2)#'
event <- rep(FALSE, nrow(data))

model <- fit(hcd_kswin(), data)
detection <- detect(model, data)
print(detection[(detection$event),])

grf <- har_plot(model, data$serie1, detection)
grf <- grf + ylab("value")
grf <- grf

plot(grf)

}
