% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hcp_garch.R
\name{hcp_garch}
\alias{hcp_garch}
\title{Change Finder using GARCH}
\usage{
hcp_garch(w = 30)
}
\arguments{
\item{w}{Sliding window size}
}
\value{
hcp_garch object
}
\description{
Change Finder using GARCH
}
\examples{
detector <- harbinger()
}
