% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hmu_pca.R
\name{hmu_pca}
\alias{hmu_pca}
\title{Multivariate anomaly detector using PCA}
\usage{
hmu_pca()
}
\value{
hmu_pca object
}
\description{
Multivariate anomaly detector using PCA
}
\examples{
detector <- hmu_pca()
}
