% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/united.R
\docType{data}
\name{mit_bih_V1}
\alias{mit_bih_V1}
\title{MIT-BIH Arrhythmia Database – V1 Lead}
\format{
A list of time series from the V1 sensor of the MIT-BIH Arrhythmia Database.
}
\source{
\doi{10.1109/51.932724}
}
\usage{
data(mit_bih_V1)
}
\description{
Data collection with real-world time-series.
MIT-BIH Arrhythmia Database (MIT-BIH).
See \href{https://github.com/cefet-rj-dal/united}{cefet-rj-dal/united}
for detailed guidance on using this package and the other datasets available in it.
Labels available? Yes
}
\details{
This package ships a mini version of the dataset. Use loadfulldata() to download and load the full dataset from the URL stored in attr(url).
}
\examples{
data(mit_bih_V1)
data <- mit_bih_V1[[1]]
series <- data$value
}
\references{
MIT-BIH Arrhythmia Database (MIT-BIH). See also: Moody, G. B., & Mark, R. G. (2001).
The impact of the MIT-BIH Arrhythmia Database. IEEE Engineering in Medicine and Biology Magazine, 20(3), 45–50.
}
\keyword{datasets}
