% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/united.R
\docType{data}
\name{nab_artificialWithAnomaly}
\alias{nab_artificialWithAnomaly}
\title{Numenta Anomaly Benchmark (NAB) – artificialWithAnomaly}
\format{
A list of time series.
}
\source{
\href{https://github.com/numenta/NAB/tree/master/data}{Numenta Anomaly Benchmark (NAB) Dataset}
}
\usage{
data(nab_artificialWithAnomaly)
}
\description{
Synthetic time series with injected anomalies from the Numenta
Anomaly Benchmark (NAB). Designed for evaluating anomaly detection
algorithms under controlled conditions. Labels available: Yes.
See \href{https://github.com/cefet-rj-dal/united}{cefet-rj-dal/united}
for detailed guidance on using this package and the other datasets available in it.
Labels available? Yes
}
\details{
This package ships a mini version of the dataset. Use loadfulldata() to download and load the full dataset from the URL stored in attr(url).
}
\examples{
data(nab_artificialWithAnomaly)
s <- nab_artificialWithAnomaly[[1]]
plot(ts(s$value), main = names(nab_artificialWithAnomaly)[1])
}
\references{
Lavin, A., & Ahmad, S. (2015). Evaluating real-time anomaly
detection algorithms – the Numenta Anomaly Benchmark. 2015 IEEE 14th
International Conference on Machine Learning and Applications (ICMLA).
}
\keyword{datasets}
