% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/united.R
\docType{data}
\name{nab_realKnownCause}
\alias{nab_realKnownCause}
\title{Numenta Anomaly Benchmark (NAB) realKnownCause}
\format{
A list of time series.
}
\source{
\href{https://github.com/numenta/NAB/tree/master/data}{Numenta Anomaly Benchmark (NAB) Dataset}
}
\usage{
data(nab_realKnownCause)
}
\description{
Data collection with real-world time-series.
Real data with anomalies
As part of the Numenta Anomaly Benchmark (NAB), this dataset contains
time series with real and synthetic data. The real data comes from network
monitoring and cloud computing. On the other hand, synthetic data simulate
series with or without anomalies.
See \href{https://github.com/cefet-rj-dal/united}{cefet-rj-dal/united}
for detailed guidance on using this package and the other datasets available in it.
Labels available? Yes
}
\details{
This package ships a mini version of the dataset. Use loadfulldata() to download and load the full dataset from the URL stored in attr(url).
}
\examples{
data(nab_realKnownCause)
nab_grp <- nab_realKnownCause[[1]]
serie <- nab_grp[[1]]
}
\references{
Lavin, A., & Ahmad, S. (2015). Evaluating real-time anomaly
detection algorithms – the Numenta Anomaly Benchmark. 2015 IEEE 14th
International Conference on Machine Learning and Applications (ICMLA).
}
\keyword{datasets}
