% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/har_outliers.R
\name{outliers.boxplot}
\alias{outliers.boxplot}
\title{Detect outliers in a dataset using the boxplot method}
\usage{
outliers.boxplot(data, alpha = 1.5)
}
\arguments{
\item{alpha}{}
}
\value{
The function returns a boolean array with true values for values considered to be outliers
}
\description{
The function receives a dataset "data" and an "alpha" parameter (default 1.5), which controls the upper limit of the boxplot
}
\details{
The function calculates the quartiles of the dataset, and then calculates the upper and lower bounds of the boxplot using the alpha parameter
}
