% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/har_plot.R
\name{plot.harbinger}
\alias{plot.harbinger}
\title{Plot event detection on a time series}
\usage{
plot.harbinger(
  obj,
  serie,
  detection,
  event = NULL,
  mark.cp = TRUE,
  ylim = NULL,
  idx = NULL
)
}
\arguments{
\item{idx}{}
}
\value{
A line graph with dots, where the time series is plotted on the y-axis and the time index is plotted on the x-axis
}
\description{
It accepts as input an object, a time series, a data.frame of events, a parameter to mark the detected change points, a threshold for the y-axis and an index for the time series
}
\details{
The function creates a data.frame that contains the time series, the event detection results (true positives, false positives, false negatives) and colors for each of the detected events. Then, a line graph is created for the time series, with colored dots for each detected event
}
