% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dal_main.R
\name{register_log}
\alias{register_log}
\title{Register log details}
\usage{
register_log(obj, msg, ref)
}
\arguments{
\item{obj}{object: .}

\item{msg}{string: a message to the log.}

\item{ref}{.}
}
\description{
Register log details
}
