% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/har_soft_evaluation.R
\name{soft_evaluation}
\alias{soft_evaluation}
\title{Harbinger}
\usage{
soft_evaluation(sw = 15)
}
\value{
Harbinger object
}
\description{
Ancestor class for time series event detection
}
\details{
The Harbinger class establishes the basic interface for time series event detection.
Each method should be implemented in a descendant class of Harbinger
}
\examples{
detector <- harbinger()
}
