% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/har_online_sources.R
\name{source_info}
\alias{source_info}
\title{Retrieve source metadata}
\usage{
source_info(obj, ...)
}
\arguments{
\item{obj}{Source object.}

\item{...}{Additional arguments passed to methods.}
}
\value{
A named list with source metadata.
}
\description{
Returns descriptive metadata about a Harbinger source object.
}
