% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trans_xsax.R
\name{trans_xsax}
\alias{trans_xsax}
\title{XSAX}
\usage{
trans_xsax(alpha)
}
\arguments{
\item{alpha}{alphabet}
}
\value{
obj
}
\description{
XSAX
}
\examples{
library(daltoolbox)
vector <- 1:52
model <- trans_xsax(alpha = 52)
model <- fit(model, vector)
xvector <- transform(model, vector)
print(xvector)
}
