% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ts_preprocessing.R
\name{ts_an}
\alias{ts_an}
\title{Time Series Adaptive Normalization}
\usage{
ts_an(remove_outliers = TRUE, nw = 0)
}
\arguments{
\item{remove_outliers}{logical: if TRUE (default) outliers will be removed.}

\item{nw}{integer: .}
}
\value{
a \code{ts_an} object.
}
\description{
Transform data to a common scale while taking into account the
changes in the statistical properties of the data over time.
}
