% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ts_preprocessing.R
\name{ts_diff}
\alias{ts_diff}
\title{Time Series Diff}
\usage{
ts_diff(remove_outliers = TRUE)
}
\arguments{
\item{remove_outliers}{logical: if TRUE (default) outliers will be removed.}
}
\value{
a \code{ts_diff} object.
}
\description{
Time Series Diff
}
