% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ts_elm.R
\name{ts_elm}
\alias{ts_elm}
\title{Time Series Extreme Learning Machine (ELM)}
\usage{
ts_elm(preprocess = NA, input_size = NA, nhid = NA, actfun = "purelin")
}
\arguments{
\item{actfun}{string: defines the type to use, possible values: 'sig',
'radbas', 'tribas', 'relu', 'purelin' (default).}
}
\value{
a \code{ts_elm} object.
}
\description{
Machine learning technique used for time series forecasting.
ELM is a type of feedforward neural network that uses a single hidden layer
of randomly generated neurons.
}
