% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ts_preprocessing.R
\name{ts_gminmax}
\alias{ts_gminmax}
\title{Time Series Global Min-Max}
\usage{
ts_gminmax(remove_outliers = TRUE)
}
\arguments{
\item{remove_outliers}{logical: if TRUE (default) outliers will be removed.}
}
\value{
a \code{ts_gminmax} object.
}
\description{
Rescales data, so the minimum value is mapped to 0 and the
maximum value is mapped to 1.
}
