% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ts_preprocessing.R
\name{ts_normalize}
\alias{ts_normalize}
\title{Time Series Normalize}
\usage{
ts_normalize(remove_outliers = TRUE)
}
\arguments{
\item{remove_outliers}{logical: if TRUE (default) outliers will be removed.}
}
\value{
a \code{ts_normalize} object.
}
\description{
Transform data to a common scale, facilitating comparisons and
analysis.
}
