% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ts_data.R
\name{ts_projection}
\alias{ts_projection}
\title{Time Series Projection}
\usage{
ts_projection(ts)
}
\arguments{
\item{ts}{matrix or data.frame containing the time series.}
}
\value{
a \code{ts_projection} object.
}
\description{
Time Series Projection
}
