% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ts_data.R
\name{ts_sample}
\alias{ts_sample}
\title{Time Series Sample}
\usage{
ts_sample(ts, test_size = 1, offset = 0)
}
\arguments{
\item{ts}{time series data.}

\item{test_size}{integer: size of test data (default = 1).}

\item{offset}{integer: starting point (default = 0).}
}
\description{
Time Series Sample
}
