% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/united.R
\docType{data}
\name{ucr_power_demand}
\alias{ucr_power_demand}
\title{UCR Anomaly Archive – Italian Power Demand}
\format{
A list of time series.
}
\source{
\href{https://paperswithcode.com/dataset/ucr-anomaly-archive/}{UCR Anomaly Archive}
}
\usage{
data(ucr_power_demand)
}
\description{
Data collection with real-world time-series.
Real power demand time series with labeled anomalous intervals.
See \href{https://github.com/cefet-rj-dal/united}{cefet-rj-dal/united}
for detailed guidance on using this package and the other datasets available in it.
Labels available? Yes
}
\details{
This package ships a mini version of the dataset. Use loadfulldata() to download and load the full dataset from the URL stored in attr(url).
}
\examples{
data(ucr_power_demand)
s <- ucr_power_demand[[1]]
summary(s$value)
}
\references{
UCR Time Series Anomaly Archive. See also: Chandola, V., Banerjee, A., & Kumar, V. (2009).
Anomaly detection: A survey. ACM Computing Surveys, 41(3), 1–58.
}
\keyword{datasets}
