% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hanr_garch.R
\name{detect.hanr_garch}
\alias{detect.hanr_garch}
\title{Performs anomaly event detection in the time series using the GARCH model}
\usage{
\method{detect}{hanr_garch}(obj, serie, ...)
}
\arguments{
\item{obj}{detector}

\item{serie}{time series}

\item{...}{optional arguments.}
}
\value{
A dataframe with information about the detected anomalous points
}
\description{
Takes as input a "Harbinger" object and a time series
}
