% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/harbinger.R
\name{detect.harbinger}
\alias{detect.harbinger}
\title{Generic outline for a time series event detection function}
\usage{
\method{detect}{harbinger}(obj, serie)
}
\arguments{
\item{serie}{}
}
\value{
A table with information about detecting events in the time series
}
\description{
Takes as input a "Harbinger" object and a time series
}
\details{
The "idx" column represents the index of the observation in the time series, the "event" column indicates whether or not the event occurred in the observation (it is filled with "TRUE" or "FALSE") and the "type" column indicates the type of event detected (if any)
}
