% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hcp_cf_ets.R
\name{detect.hcp_cf_ets}
\alias{detect.hcp_cf_ets}
\title{Detect changes in a time series}
\usage{
\method{detect}{hcp_cf_ets}(obj, serie)
}
\arguments{
\item{serie}{}
}
\value{
A data frame with information about the events detected in the time series, including the index of the event, whether it is an outlier or change point and the type of event
}
\description{
The function takes an object of the Harbinger class and a time series as parameters
}
\details{
Detection is done using the ETS model to fit the time series and statistical methods to detect change points and outliers
}
