% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hcp_scp.R
\name{detect.hcp_scp}
\alias{detect.hcp_scp}
\title{Seminal change Point Detector}
\usage{
\method{detect}{hcp_scp}(obj, serie, ...)
}
\arguments{
\item{obj}{detector}

\item{serie}{time series}

\item{...}{optional arguments.}
}
\value{
A dataframe with information about the detected anomalous points
}
\description{
Takes as input a "Harbinger" object and a time series
}
