% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/harbinger.R
\name{detect}
\alias{detect}
\title{Detect events in time series}
\usage{
detect(obj, ...)
}
\arguments{
\item{obj}{harbinger object}

\item{...}{optional arguments.}
}
\value{
a data frame with the index of observations and if they are identified or not as an event, and their type
}
\description{
Event detection using a fitted Harbinger model
}
\examples{
#See examples of detectors for anomalies, change points, and motifs
#at https://cefet-rj-dal.github.io/harbinger
}
