% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/harbinger.R
\name{detect}
\alias{detect}
\title{Detect events in time series}
\usage{
detect(obj, ...)
}
\arguments{
\item{obj}{harbinger object}

\item{...}{optional arguments./ further arguments passed to or from other methods.}
}
\value{
a harbinger object with model details
}
\description{
Detect event using a Harbinger model for event detection
}
\examples{
data(har_examples)
dataset <- har_examples[[1]]
detector <- harbinger()
detector <- detect(detector, dataset$serie)
}
