% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/har_eval.R
\name{evaluate.har_eval}
\alias{evaluate.har_eval}
\title{Evaluates the model based on a confusion matrix}
\usage{
\method{evaluate}{har_eval}(obj, detection, event, ...)
}
\arguments{
\item{obj}{detector}

\item{detection}{detected observations}

\item{event}{labeled events}

\item{...}{optional arguments.}
}
\value{
An object of the "hard evaluation" class updated with the calculated metrics
}
\description{
It takes three arguments: obj: an object to store the evaluation results; detection: a boolean vector that indicates whether or not the model detected an event; event: a boolean vector that indicates whether the event occurred or not
}
\examples{
detector <- harbinger()
}
