% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/har_eval_soft.R
\name{evaluate.har_eval_soft}
\alias{evaluate.har_eval_soft}
\title{Evaluate the quality of an event detection model against a reference dataset}
\usage{
\method{evaluate}{har_eval_soft}(obj, detection, event, ...)
}
\arguments{
\item{obj}{detector}

\item{detection}{detected observations}

\item{event}{labeled events}

\item{...}{optional arguments.}
}
\value{
A list of performance measures for detection
}
\description{
It receives as input an object, a "detection" matrix with the detections made by the model and an "event" matrix with the true events
}
\examples{
detector <- harbinger()
}
