% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{examples_anomalies}
\alias{examples_anomalies}
\title{Time series for anomaly detection}
\format{
A list of time series for anomaly detection.
}
\source{
\href{https://github.com/cefet-rj-dal/harbinger}{Harbinger package}
}
\usage{
data(examples_anomalies)
}
\description{
A list of time series for anomaly detection
\itemize{
\item simple: a simple synthetic time series
\item contextual: a contextual synthetic time series
\item simple: a trend synthetic time series
\item trend: a simple synthetic time series
\item multiple: a multiple anomalies synthetic time series
\item sequence: a sequence synthetic time series
\item tt: a train-test synthetic time series
\item tt_warped: a warped train-test synthetic time series
}
#'
}
\examples{
data(examples_anomalies)
serie <- examples_anomalies$simple
}
\references{
\href{https://github.com/cefet-rj-dal/harbinger}{Harbinger package}
}
\keyword{datasets}
