% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hanr_ml.R
\name{fit.hanr_ml}
\alias{fit.hanr_ml}
\title{Fits a detection method using a regression model}
\usage{
\method{fit}{hanr_ml}(obj, serie, ...)
}
\arguments{
\item{obj}{detector}

\item{serie}{time series}

\item{...}{optional arguments.}
}
\value{
Adjusted model
}
\description{
Takes as input a "Harbinger" object and a time series
}
\examples{
detector <- harbinger()
}
