% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/harbinger.R
\name{fit.harbinger}
\alias{fit.harbinger}
\title{Fit a general detector}
\usage{
\method{fit}{harbinger}(obj, serie, ...)
}
\arguments{
\item{obj}{detector}

\item{serie}{time series}

\item{...}{optional arguments.}
}
\value{
The "Harbinger" object without doing any processing or tuning of the model
}
\description{
Takes as input a "Harbinger" object and a time series
}
\examples{
detector <- harbinger()
}
