% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dal_main.R, R/harbinger.R
\name{fit}
\alias{fit}
\title{Fit a model for event detection}
\usage{
fit(obj, ...)

fit(obj, ...)
}
\arguments{
\item{obj}{harbinger object}

\item{...}{optional arguments./ further arguments passed to or from other methods.}
}
\value{
a harbinger object with model details
}
\description{
Basic ancestor function for build model for event detection
}
\details{
The fit function builds a model for time series event detection.
For some methods, the model is not needed to be build, so the function do nothing.
}
\examples{
data(har_examples)
dataset <- har_examples[[1]]
detector <- harbinger()
detector <- fit(detector, dataset$serie)
}
