% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/han_autoencoder.R
\name{han_autoencoder}
\alias{han_autoencoder}
\title{Anomaly detector using autoencoder}
\usage{
han_autoencoder(input_size, encode_size)
}
\arguments{
\item{input_size}{Establish the input size for the autoencoder anomaly detector. It is the size of the output also.}

\item{encode_size}{The encode size for the autoencoder.}
}
\value{
han_autoencoder object
histogram based method to detect anomalies in time series. Bins with smaller amount of observations are considered anomalies. Values below first bin or above last bin are also considered anomalies.>.
}
\description{
Anomaly detector using autoencoder
}
\examples{
# setting up time series regression model
#Use the same example of hanr_fbiad changing the constructor to:
model <- han_autoencoder(3,1)
}
