% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/han_autoencoder.R
\name{han_autoencoder}
\alias{han_autoencoder}
\title{Anomaly detector using autoencoder}
\usage{
han_autoencoder(input_size, encode_size, encoderclass = autoenc_ed, ...)
}
\arguments{
\item{input_size}{Establish the input size for the autoencoder anomaly detector. It is the size of the output also.}

\item{encode_size}{The encode size for the autoencoder.}

\item{encoderclass}{The class of daltoolbox encoder-decoder.}

\item{...}{optional arguments for encoder-decoder class.}
}
\value{
han_autoencoder object
histogram based method to detect anomalies in time series. Bins with smaller amount of observations are considered anomalies. Values below first bin or above last bin are also considered anomalies.>.
}
\description{
Anomaly detector using autoencoder
}
\examples{
#See an example of using `autoenc_ed` at this
#https://github.com/cefet-rj-dal/harbinger/blob/master/anomalies/han_autoenc_ed
}
