% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hanc_ml.R
\name{hanc_ml}
\alias{hanc_ml}
\title{Anomaly detector based on machine learning classification}
\usage{
hanc_ml(model, tune = NULL, alpha = 1.5)
}
\arguments{
\item{model}{DAL Classification Model}

\item{tune}{DAL Tune Model}

\item{alpha}{Outliers threshold}
}
\value{
hanc_ml object
}
\description{
Anomaly detector based on machine learning classification
}
\examples{
detector <- harbinger()
}
