% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hanr_fbiad.R
\name{hanr_fbiad}
\alias{hanr_fbiad}
\title{Anomaly detector using FBIAD}
\usage{
hanr_fbiad(sw = 30, alpha = 1.5)
}
\arguments{
\item{sw}{Window size for FBIAD}

\item{alpha}{Threshold for outliers}
}
\value{
hanr_fbiad object
}
\description{
Anomaly detector using FBIAD
}
\examples{
detector <- harbinger()
}
