% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hanr_ml.R
\name{hanr_ml}
\alias{hanr_ml}
\title{Anomaly detector using machine learning regression}
\usage{
hanr_ml(model, tune = NULL, sw_size = 15, alpha = 1.5)
}
\arguments{
\item{model}{DAL Toolbox regression model}

\item{tune}{DAL Toolbox tunel}

\item{sw_size}{Sliding window size}

\item{alpha}{Threshold for outliers}
}
\value{
hanr_ml object
}
\description{
Anomaly detector using machine learning regression
}
\examples{
detector <- harbinger()
}
