% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hanr_red.R
\name{hanr_red}
\alias{hanr_red}
\title{Anomaly and change point detector using RED}
\usage{
hanr_red(
  noise = 0.001,
  trials = 5,
  red_cp = TRUE,
  volatility_cp = TRUE,
  trend_cp = TRUE
)
}
\arguments{
\item{noise}{nosie}

\item{trials}{trials}

\item{red_cp}{red change point}

\item{volatility_cp}{volatility change point}

\item{trend_cp}{trend change point}
}
\value{
\code{hanr_emd} object
}
\description{
Anomaly and change point detection using RED
The RED model adjusts to the time series. Observations distant from the model are labeled as anomalies.
It wraps the EMD model presented in the hht library.
}
\examples{
library(daltoolbox)

#loading the example database
data(har_examples)

#Using example 1
dataset <- har_examples$example1
head(dataset)

# setting up time series emd detector
model <- hanr_emd()

# fitting the model
model <- fit(model, dataset$serie)

detection <- detect(model, dataset$serie)

# filtering detected events
print(detection[(detection$event),])

}
