% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hanr_remd.R
\name{hanr_remd}
\alias{hanr_remd}
\title{Anomaly detector using REMD}
\usage{
hanr_remd(noise = 0.1, trials = 5)
}
\arguments{
\item{noise}{nosie}

\item{trials}{trials}
}
\value{
\code{hanr_remd} object
}
\description{
Anomaly detection using REMD
The EMD model adjusts to the time series. Observations distant from the model are labeled as anomalies.
It wraps the EMD model presented in the forecast library.
}
\examples{
library(daltoolbox)

#loading the example database
data(examples_anomalies)

#Using simple example
dataset <- examples_anomalies$simple
head(dataset)

# setting up time series emd detector
model <- hanr_remd()

# fitting the model
model <- fit(model, dataset$serie)

detection <- detect(model, dataset$serie)

# filtering detected events
print(detection[(detection$event),])

}
