% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hanr_wavelet.R
\name{hanr_wavelet}
\alias{hanr_wavelet}
\title{Anomaly detector using WAVELET.}
\usage{
hanr_wavelet()
}
\value{
\code{hanr_wavelet} object
}
\description{
Anomaly detection using WAVELET
The function uses the discrete Wavelet transform to detect anomalies through the frequency component.
Observations in the frequency component that are outside the tolerance (2 * 2,698 standard deviations) are considered anomalies.
}
\examples{
library(daltoolbox)

#loading the example database
data(har_examples)

#Using example 1
dataset <- har_examples$example1
head(dataset)

# setting up time series regression model
model <- hanr_wavelet()

# fitting the model
model <- fit(model, dataset$serie)

detection <- detect(model, dataset$serie)

# filtering detected events
print(detection[(detection$event),])

}
