% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/har_ensemble.R
\name{har_ensemble}
\alias{har_ensemble}
\title{Harbinger Ensemble}
\usage{
har_ensemble(...)
}
\arguments{
\item{...}{list of detectors}
}
\value{
Harbinger object
}
\description{
Ensemble detector
}
\examples{
library(daltoolbox)

#loading the example database
data(examples_anomalies)

#Using simple example
dataset <- examples_anomalies$simple
head(dataset)

# setting up time series emd detector
model <- har_ensemble(hanr_arima(), hanr_arima(), hanr_arima())
#model <- har_ensemble(hanr_fbiad(), hanr_arima(), hanr_emd())

# fitting the model
model <- fit(model, dataset$serie)

detection <- detect(model, dataset$serie)

# filtering detected events
print(detection[(detection$event),])
}
