% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/har_ensemble.R
\name{har_ensemble}
\alias{har_ensemble}
\title{Harbinger Ensemble}
\usage{
har_ensemble()
}
\value{
Harbinger object
}
\description{
Ensemble detector
}
\examples{
# See ?hanc_ml for an example of anomaly detection using machine learning classification
# See ?hanr_arima for an example of anomaly detection using ARIMA
# See ?hanr_fbiad for an example of anomaly detection using FBIAD
# See ?hanr_garch for an example of anomaly detection using GARCH
# See ?hanr_kmeans for an example of anomaly detection using kmeans clustering
# See ?hanr_ml for an example of anomaly detection using machine learning regression
# See ?hanr_cf_arima for an example of change point detection using ARIMA
# See ?hanr_cf_ets for an example of change point detection using ETS
# See ?hanr_cf_lr for an example of change point detection using linear regression
# See ?hanr_cf_garch for an example of change point detection using GARCH
# See ?hanr_cf_scp for an example of change point detection using the seminal algorithm
# See ?hmo_sax for an example of motif discovery using SAX
# See ?hmu_pca for an example of anomaly detection in multivariate time series using PCA
}
