% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/har_eval_soft.R
\name{har_eval_soft}
\alias{har_eval_soft}
\title{Evaluate the quality of an event detection model against a reference dataset}
\usage{
har_eval_soft(sw = 15)
}
\arguments{
\item{sw}{tolerance window size}
}
\value{
Harbinger object
}
\description{
It receives as input an object, a "detection" matrix with the detections made by the model and an "event" matrix with the true events
}
\examples{
detector <- harbinger()
}
