% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/har_eval_soft.R
\name{har_eval_soft}
\alias{har_eval_soft}
\title{Harbinger}
\usage{
har_eval_soft(sw = 15)
}
\arguments{
\item{sw}{tolerance window size}
}
\value{
Harbinger object
}
\description{
Ancestor class for time series event detection
}
\examples{
detector <- harbinger()
}
