% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{har_examples_multi}
\alias{har_examples_multi}
\title{Synthetic time series for event detection}
\format{
A list of multivariate time series.
}
\source{
\href{https://github.com/cefet-rj-dal/harbinger}{Harbinger package}
}
\usage{
data(har_examples_multi)
}
\description{
A list of multivariate time series for event detection
\itemize{
\item example1: an single multivariate anomaly
}
#'
}
\examples{
data(har_examples_multi)
serie <- har_examples_multi$example1
}
\references{
\href{https://github.com/cefet-rj-dal/harbinger}{Harbinger package}
}
\keyword{datasets}
