% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/har_plot.R
\name{har_plot}
\alias{har_plot}
\title{Plot event detection on a time series}
\usage{
har_plot(
  obj,
  serie,
  detection,
  event = NULL,
  mark.cp = TRUE,
  ylim = NULL,
  idx = NULL
)
}
\arguments{
\item{obj}{detector}

\item{serie}{time series}

\item{detection}{detections}

\item{event}{events}

\item{mark.cp}{mark change point}

\item{ylim}{limits for y-axis}

\item{idx}{labels for x observations}
}
\value{
A line graph with dots, where the time series is plotted on the y-axis and the time index is plotted on the x-axis
}
\description{
It accepts as input an object, a time series, a data.frame of events, a parameter to mark the detected change points, a threshold for the y-axis and an index for the time series
}
\examples{
detector <- harbinger()
}
