% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/harbinger.R
\name{harbinger}
\alias{harbinger}
\title{Harbinger}
\usage{
harbinger()
}
\value{
Harbinger object
}
\description{
Ancestor class for time series event detection
}
\examples{
#See examples of detectors for anomalies, change points, and motifs
#at https://cefet-rj-dal.github.io/harbinger
}
