% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/harbinger.R
\name{harbinger}
\alias{harbinger}
\title{Harbinger}
\usage{
harbinger()
}
\value{
Harbinger object
}
\description{
Ancestor class for time series event detection
}
\details{
The Harbinger class establishes the basic interface for time series event detection.
Each method should be implemented in a descendant class of Harbinger
}
\examples{
detector <- harbinger()
}
