% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/harbinger_utils.R
\name{harutils}
\alias{harutils}
\title{Harbinger Utils}
\usage{
harutils()
}
\value{
Harbinger Utils
}
\description{
Utility class that contains major distance measures,
threshold limits, and outliers grouping functions
}
\examples{
# See ?hanc_ml for an example of anomaly detection using machine learning classification
}
