% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hcp_cf_arima.R
\name{hcp_cf_arima}
\alias{hcp_cf_arima}
\title{Change Finder using ARIMA}
\usage{
hcp_cf_arima(w = NULL, alpha = 1.5)
}
\arguments{
\item{w}{Sliding window size}

\item{alpha}{Threshold for outliers}
}
\value{
hcp_cf_arima object
}
\description{
Change Finder using ARIMA
}
\examples{
detector <- harbinger()
}
