% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hcp_cf_ets.R
\name{hcp_cf_ets}
\alias{hcp_cf_ets}
\title{Change Finder using ETS}
\usage{
hcp_cf_ets(w = 7, alpha = 1.5)
}
\arguments{
\item{w}{Sliding window size}

\item{alpha}{Threshold for outliers}
}
\value{
hcp_cf_arima object
}
\description{
Change Finder using ETS
}
\examples{
detector <- harbinger()
}
