% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hcp_cf_lr.R
\name{hcp_cf_lr}
\alias{hcp_cf_lr}
\title{Change Finder using linear regression}
\usage{
hcp_cf_lr(w = 30)
}
\arguments{
\item{w}{Sliding window size}
}
\value{
hcp_cf_lr object
}
\description{
Change Finder using linear regression
}
\examples{
detector <- harbinger()
}
