% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hcp_scp.R
\name{hcp_scp}
\alias{hcp_scp}
\title{Seminal change point}
\usage{
hcp_scp(sw = 30, alpha = 1.5)
}
\arguments{
\item{sw}{Sliding window size}

\item{alpha}{Threshold for outliers}
}
\value{
hcp_scp object
}
\description{
Seminal change point
}
\examples{
detector <- harbinger()
}
