% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hmo_sax.R
\name{hmo_sax}
\alias{hmo_sax}
\title{Motif discovery using SAX}
\usage{
hmo_sax(a, w, qtd)
}
\arguments{
\item{a}{alphabet size}

\item{w}{word size}

\item{qtd}{number of occurrences to be classified as motifs}
}
\value{
hmo_sax object
}
\description{
Motif discovery using SAX
}
\examples{
detector <- harbinger()
}
