% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trans_sax.R
\name{trans_sax}
\alias{trans_sax}
\title{SAX}
\usage{
trans_sax(alpha)
}
\arguments{
\item{alpha}{alphabet}
}
\value{
obj
}
\description{
SAX
}
\examples{
library(daltoolbox)
vector <- 1:52
model <- trans_sax(alpha = 26)
model <- fit(model, vector)
xvector <- transform(model, vector)
print(xvector)
}
