% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dal_main.R
\name{transform}
\alias{transform}
\title{Transform}
\usage{
transform(obj, ...)
}
\arguments{
\item{obj}{object: .}

\item{...}{optional arguments./ further arguments passed to or from other methods.}
}
\description{
Transform
}
